agents/story_processor.py
from azure.ai.textanalytics import TextAnalyticsClient
from azure.core.credentials import AzureKeyCredential
import json
import re
import logging

logger = logging.getLogger('kidsbook')

class StoryProcessor:
    def __init__(self, azure_config_path="azure_config.json"):
        self.azure_config_path = azure_config_path
        # Add any initialization that depends on the config here

    def load_azure_config(self, path):
        with open(path) as config_file:
            return json.load(config_file)

    def create_text_analytics_client(self):
        endpoint = self.azure_config['endpoint']
        key = self.azure_config['key']
        return TextAnalyticsClient(endpoint=endpoint, credential=AzureKeyCredential(key))

    def filter_content(self, story):
        # Implement content filtering logic here
        # This should include checks for appropriateness and adherence to responsible AI principles
        return story  # Placeholder for filtered story

    def process_story(self, story):
        filtered_story = self.filter_content(story)
        # Additional processing logic can be added here
        return filtered_story

    def analyze_story(self, story):
        processed_story = self.process_story(story)
        # Call Azure AI services for further analysis if needed
        return processed_story

    def process(self, final_story: str, illustrations: str, illustrator_prompt: str):
        """
        Combines the final edited story and illustration information.
        This is a synchronous, CPU-bound operation, so it is run in a thread.
        """
        composite = (
            f"{final_story}\n\n"
            f"Illustration URL: {illustrations}\n"
            f"Prompt used: {illustrator_prompt}"
        )
        logger.info("StoryProcessor: process completed")
        return composite